requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
apscheduler==3.10.4
pydantic==2.5.2
rich==13.7.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
matplotlib==3.8.2